refactor(imports): update schema imports  to new entities locations
