+++ apps/backend/src/health/health.service.ts
@@ -1,6 +1,6 @@
 import { Injectable, Logger } from '@nestjs/common';
 import { DatabaseService } from '../db/database.service.js';
-import { HelloWorld, helloWorldTable } from './hello-world.schema.js';
+import { HelloWorld, helloWorldTable } from '../entities/hello-world/index.js';
 
 @Injectable()
 export class HealthService {
